config controlled ixpmgr versioning (#5) rename v57 to v5, rename v37 to v3 (#6)
