+++ src/django_ixpmgr/v3/__init__.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
 
 class IxpManagerAppConfig(AppConfig):
-    name = "django_ixpmgr.v57"
-    # Must be unique, will conflict if two versions are used
+    name = "django_ixpmgr.v3"
     label = "django_ixpmgr"
