+++ src/django_ixpmgr/v5/__init__.py
@@ -1,5 +1,6 @@
 from django.apps import AppConfig
 
 class IxpManagerAppConfig(AppConfig):
-    name = "django_ixpmgr.v37"
+    name = "django_ixpmgr.v5"
+    # Must be unique, will conflict if two versions are used
     label = "django_ixpmgr"
